aad-machine-to-machine/requirements.txt
blacksheep
msal
uvicorn
python-dotenv
httpx
